refactor end level function to its own class
