+++ lib/index.js
@@ -9,6 +9,7 @@ var Game = require("./game");
 var Club = require("./club")
 var calculateScore = require("./scoreCalculator");
 var endLevel = require("./end-level");
+var transitionToNextLevel = require("./transitionLevel");
 
 $(document).ready(function() {
     $(".a-ball-color").on('click', function(ev) {
@@ -55,23 +56,16 @@ $(document).ready(function() {
           };
 
           function endLevel() {
-              var currentLevel = game.currentLevel.number
-              var currentPar = game.currentLevel.par
-              var score = calculateScore(currentPar, strokeCounter)
               if(game.currentLevel.number < game.lastLevel) {
-                  game.updateLevel(context);
-                  var nextLevel = game.currentLevel.number
-                  var nextPar = game.currentLevel.par
-                  strokeCounter = 0
-                  domManipulator.transitionLevel(currentLevel, currentPar, nextLevel, nextPar, score)
+                  transitionToNextLevel(context, game, strokeCounter)
                   requestAnimationFrame(gameLoop);
               } else {
                   requestAnimationFrame(endGame);
               }
+              strokeCounter = 0
           }
 
           function endGame() {
-            //
           }
 
         canvas.addEventListener('mousedown', function(event) {
